+++ packages/orchestrator-store/src/types.d.ts
@@ -1,0 +1,2 @@
+declare module 'hapi__catbox';
+declare module 'hapi__shot';